test: combine policy e2e tests & test inputs
