+++ edc-tests/e2e/policy-tests/src/test/java/org/eclipse/tractusx/edc/tests/policy/PolicyDefinitionEndToEndTest.java
@@ -25,6 +25,7 @@ import io.restassured.http.ContentType;
 import io.restassured.response.Response;
 import jakarta.json.Json;
 import jakarta.json.JsonObject;
+import jakarta.json.JsonValue;
 import org.eclipse.edc.junit.annotations.EndToEndTest;
 import org.eclipse.edc.junit.extensions.RuntimeExtension;
 import org.eclipse.edc.policy.model.Operator;
@@ -41,13 +42,17 @@ import org.junit.jupiter.params.provider.ArgumentsSource;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 
+import static com.apicatalog.jsonld.lang.Keywords.ID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.CONTEXT;
 import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.TYPE;
+import static org.eclipse.edc.spi.constants.CoreConstants.EDC_CONNECTOR_MANAGEMENT_CONTEXT_V2;
 import static org.eclipse.tractusx.edc.cx.CxJsonLdExtension.CX_POLICY_2025_09_CONTEXT;
 import static org.eclipse.tractusx.edc.tests.TestRuntimeConfiguration.CONSUMER_BPN;
 import static org.eclipse.tractusx.edc.tests.TestRuntimeConfiguration.CONSUMER_DID;
@@ -64,6 +69,7 @@ import static org.eclipse.tractusx.edc.tests.runtimes.Runtimes.pgRuntime;
 
 @EndToEndTest
 public class PolicyDefinitionEndToEndTest {
+
     private static final TransferParticipant CONSUMER = TransferParticipant.Builder.newInstance()
             .name(CONSUMER_NAME)
             .id(CONSUMER_DID)
@@ -88,114 +94,113 @@ public class PolicyDefinitionEndToEndTest {
     private static final RuntimeExtension PROVIDER_RUNTIME = pgRuntime(PROVIDER, POSTGRES);
 
     @DisplayName("Policy is accepted")
-    @ParameterizedTest(name = "{1}")
+    @ParameterizedTest(name = "{2}")
     @ArgumentsSource(ValidContractPolicyProvider.class)
-    void shouldAcceptValidPolicyDefinitions(JsonObject policy, String description) {
-        var response = createPolicyDefinition(policy);
+    void shouldAcceptValidPolicyDefinitions(ManagementApiVersion apiVersion, JsonObject policy, String description) {
+        var response = createPolicyDefinition(apiVersion, policy);
         assertThat(response.statusCode()).isEqualTo(200);
     }
 
-    @DisplayName("Policy is not accepted due to missing context")
-    @ParameterizedTest(name = "{1}")
-    @ArgumentsSource(InValidNamespaceContractPolicyProvider.class)
-    void shouldNotAcceptInvalidNamespacePolicyDefinitions(JsonObject policy, String description) {
-        checkForValidationFailure(policy);
-    }
-
     @DisplayName("Policy is not accepted because definition is not correct")
-    @ParameterizedTest(name = "{1}")
-    @ArgumentsSource(InValidContractPolicyProvider.class)
-    void shouldNotAcceptInvalidPolicyDefinitions(JsonObject policy, String description) {
-        checkForValidationFailure(policy);
+    @ParameterizedTest(name = "{2}")
+    @ArgumentsSource(InvalidContractPolicyProvider.class)
+    void shouldNotAcceptInvalidPolicyDefinitions(ManagementApiVersion apiVersion, JsonObject policy, String description) {
+        checkForValidationFailure(apiVersion, policy);
     }
 
-    private void checkForValidationFailure(JsonObject policy) {
-        var response = createPolicyDefinition(policy);
+    private void checkForValidationFailure(ManagementApiVersion apiVersion, JsonObject policy) {
+        var response = createPolicyDefinition(apiVersion, policy);
         assertThat(response.statusCode()).isEqualTo(400);
         assertThat(response.body().jsonPath().getString("[0].type")).isEqualTo("ValidationFailure");
     }
 
-    private abstract static class BasePolicyProvider implements ArgumentsProvider {
+    private abstract static class ApiVersionsArgumentsProvider implements ArgumentsProvider {
+        List<ManagementApiVersion> apiVersions = List.of(ManagementApiVersion.V3, ManagementApiVersion.V4);
 
-        protected final String namespace;
+        protected abstract Stream<? extends Arguments> arguments();
 
-        private BasePolicyProvider(String namespace) {
-            this.namespace = namespace;
-        }
-
+        /**
+         * Creates a new set of arguments, where every combination of arguments from the subclass
+         * is combined with management API versions v3 & v4.
+         *
+         * @param extensionContext the current extension context; never {@code null}
+         * @return a stream of arguments, where each combination of original arguments is paired with both management API versions (v3 & v4)
+         */
         @Override
         public Stream<? extends Arguments> provideArguments(ExtensionContext extensionContext) {
-            return Stream.of(
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("Membership", "active"), "use", Operator.EQ, false)), "MembershipCredential"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("FrameworkAgreement", "DataExchangeGovernance:1.0"), "use", Operator.EQ, false)), "DataExchangeGovernance use case"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("AffiliatesRegion", List.of("cx.region.all:1", "cx.region.europe:1", "cx.region.northAmerica:1")), "use", Operator.IS_ANY_OF, true)), "Affiliates Region"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("AffiliatesRegion", List.of("cx.region.europe:1")), "use", Operator.IS_ANY_OF, true)), "Affiliates Region (IS_ANY_OF, one element)"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("AffiliatesBpnl", "BPNL00000000001A"), "use", Operator.IS_ANY_OF, false)), "Affiliates BPNL"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("DataFrequency", "cx.dataFrequency.once:1"), "use", Operator.EQ, false)), "Data Frequency"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("DataUsageEndDate", "2025-06-30T14:30:00Z"), "use", Operator.EQ, false)), "Data Usage End Date"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("DataUsageEndDefinition", "cx.dataUsageEnd.unlimited:1"), "use", Operator.EQ, false)), "Data Usage End Date Definition"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("DataUsageEndDurationDays", 3), "use", Operator.EQ, false)), "Data Usage End Duration Days"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("JurisdictionLocation", "test location"), "use", Operator.EQ, false)), "Jurisdiction Location"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("JurisdictionLocationReference", "cx.location.dataConsumer:1"), "use", Operator.EQ, false)), "Jurisdiction Location Reference"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("Liability", "cx.grossNegligence:1"), "use", Operator.EQ, false)), "Liability"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("Liability", "cx.slightNegligence:1"), "use", Operator.EQ, false)), "Liability"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("Precedence", "cx.precedence.contractReference:1"), "use", Operator.EQ, false)), "Precedence"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("UsagePurpose", List.of("cx.core.legalRequirementForThirdparty:1", "cx.core.industrycore:1")), "use", Operator.IS_ANY_OF, true)), "Usage Purpose"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("VersionChanges", "cx.versionChanges.minor:1"), "use", Operator.EQ, false)), "Version Changes"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("Warranty", "cx.warranty.none:1"), "use", Operator.EQ, false)), "Warranty"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("WarrantyDefinition", "cx.warranty.contractEndDate:1"), "use", Operator.EQ, false)), "Warranty Definition"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("WarrantyDurationMonths", 3), "use", Operator.EQ, false)), "Warranty Duration Months"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("ExclusiveUsage", "cx.exclusiveUsage.dataConsumer:1"), "use", Operator.EQ, false)), "Exclusive Usage"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("ContractReference", List.of("contractReference")), "use", Operator.IS_ALL_OF, true)), "Contract reference"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("ContractTermination", "cx.data.deletion:1"), "use", Operator.EQ, false)), "ContractTermination"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("ConfidentialInformationMeasures", "cx.confidentiality.measures:1"), "use", Operator.EQ, false)), "Confidential Information Measures"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("ConfidentialInformationSharing", List.of("cx.sharing.affiliates:1")), "use", Operator.IS_ANY_OF, true)), "Confidential Information Sharing"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("BusinessPartnerGroup", List.of("Some-group")), "access", Operator.IS_ANY_OF, true)), "Business Partner Group"),
-                    Arguments.of(policyFromRules("permission", namespace,
-                            frameworkConstraint(Map.of("BusinessPartnerNumber", List.of("BPNL00000000001A")), "access", Operator.IS_ANY_OF, true)), "Business Partner Number")
-            );
+            return arguments()
+                    .flatMap(arguments -> apiVersions.stream()
+                            .map(apiVersion -> {
+                                var newParams = new ArrayList<>();
+                                newParams.add(apiVersion);
+                                newParams.addAll(Arrays.asList(arguments.get()));
+                                return Arguments.of(newParams.toArray());
+                            }));
         }
     }
 
-    private static class ValidContractPolicyProvider extends BasePolicyProvider {
-
-        private ValidContractPolicyProvider() {
-            super(CX_POLICY_2025_09_CONTEXT);
-        }
+    private static class ValidContractPolicyProvider extends ApiVersionsArgumentsProvider {
 
         @Override
-        public Stream<? extends Arguments> provideArguments(ExtensionContext extensionContext) {
-            return Stream.concat(super.provideArguments(extensionContext), Stream.of(
+        public Stream<? extends Arguments> arguments() {
+            return Stream.of(
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("Membership", "active"), "use", Operator.EQ, false)), "MembershipCredential"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("FrameworkAgreement", "DataExchangeGovernance:1.0"), "use", Operator.EQ, false)), "DataExchangeGovernance use case"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("AffiliatesRegion", List.of("cx.region.all:1", "cx.region.europe:1", "cx.region.northAmerica:1")), "use", Operator.IS_ANY_OF, true)), "Affiliates Region"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("AffiliatesRegion", List.of("cx.region.europe:1")), "use", Operator.IS_ANY_OF, true)), "Affiliates Region (IS_ANY_OF, one element)"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("AffiliatesBpnl", "BPNL00000000001A"), "use", Operator.IS_ANY_OF, true)), "Affiliates BPNL"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("DataFrequency", "cx.dataFrequency.once:1"), "use", Operator.EQ, false)), "Data Frequency"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("DataUsageEndDate", "2025-06-30T14:30:00Z"), "use", Operator.EQ, false)), "Data Usage End Date"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("DataUsageEndDefinition", "cx.dataUsageEnd.unlimited:1"), "use", Operator.EQ, false)), "Data Usage End Date Definition"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("DataUsageEndDurationDays", 3), "use", Operator.EQ, false)), "Data Usage End Duration Days"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("JurisdictionLocation", "test location"), "use", Operator.EQ, false)), "Jurisdiction Location"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("JurisdictionLocationReference", "cx.location.dataConsumer:1"), "use", Operator.EQ, false)), "Jurisdiction Location Reference"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("Liability", "cx.grossNegligence:1"), "use", Operator.EQ, false)), "Liability"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("Liability", "cx.slightNegligence:1"), "use", Operator.EQ, false)), "Liability"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("Precedence", "cx.precedence.contractReference:1"), "use", Operator.EQ, false)), "Precedence"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("UsagePurpose", List.of("cx.core.legalRequirementForThirdparty:1", "cx.core.industrycore:1")), "use", Operator.IS_ANY_OF, true)), "Usage Purpose"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("VersionChanges", "cx.versionChanges.minor:1"), "use", Operator.EQ, false)), "Version Changes"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("Warranty", "cx.warranty.none:1"), "use", Operator.EQ, false)), "Warranty"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("WarrantyDefinition", "cx.warranty.contractEndDate:1"), "use", Operator.EQ, false)), "Warranty Definition"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("WarrantyDurationMonths", 3), "use", Operator.EQ, false)), "Warranty Duration Months"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("ExclusiveUsage", "cx.exclusiveUsage.dataConsumer:1"), "use", Operator.EQ, false)), "Exclusive Usage"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("ContractReference", List.of("contractReference")), "use", Operator.IS_ALL_OF, true)), "Contract reference"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("ContractTermination", "cx.data.deletion:1"), "use", Operator.EQ, false)), "ContractTermination"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("ConfidentialInformationMeasures", "cx.confidentiality.measures:1"), "use", Operator.EQ, false)), "Confidential Information Measures"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("ConfidentialInformationSharing", List.of("cx.sharing.affiliates:1")), "use", Operator.IS_ANY_OF, true)), "Confidential Information Sharing"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("BusinessPartnerGroup", List.of("Some-group")), "access", Operator.IS_ANY_OF, true)), "Business Partner Group"),
+                    Arguments.of(policyFromRules("permission",
+                            frameworkConstraint(Map.of("BusinessPartnerNumber", List.of("BPNL00000000001A")), "access", Operator.IS_ANY_OF, true)), "Business Partner Number"),
                     Arguments.of(
                             emptyPolicy(),
                             "Empty Policy"),
                     Arguments.of(
-                            policyWithEmptyRule("access", this.namespace),
+                            policyWithEmptyRule("access"),
                             "Access policy with empty permission"),
                     Arguments.of(
                             dataUsageEndDurationDays(1),
@@ -203,74 +208,80 @@ public class PolicyDefinitionEndToEndTest {
                     Arguments.of(
                             dataUsageEndDate(Instant.now().plus(1, ChronoUnit.SECONDS).truncatedTo(ChronoUnit.SECONDS).toString()),
                             "Enforce data usage end date 1 second in the future")
-            ));
+            );
         }
     }
 
-    private static class InValidNamespaceContractPolicyProvider extends BasePolicyProvider {
-
-        private InValidNamespaceContractPolicyProvider() {
-            super("");
-        }
-    }
-
-    private static class InValidContractPolicyProvider extends BasePolicyProvider {
-
-        private InValidContractPolicyProvider() {
-            super(CX_POLICY_2025_09_CONTEXT);
-        }
+    private static class InvalidContractPolicyProvider extends ApiVersionsArgumentsProvider {
 
         @Override
-        public Stream<? extends Arguments> provideArguments(ExtensionContext extensionContext) {
+        public Stream<? extends Arguments> arguments() {
             return Stream.of(
-                    Arguments.of(policyWithEmptyRule("use", this.namespace), "Usage policy with empty permission"),
-                    Arguments.of(policyFromRules("permission", namespace,
+                    Arguments.of(policyWithEmptyRule("use"), "Usage policy with empty permission"),
+                    Arguments.of(policyFromRules("permission",
                             frameworkConstraint(Map.of("Membership", "active"), "access", Operator.EQ, false),
                             frameworkConstraint(Map.of("UsagePurpose", List.of("cx.core.industrycore:1")), "use", Operator.IS_ANY_OF, true)), "Policy with different actions types"),
-                    Arguments.of(policyFromRules("permission", namespace,
+                    Arguments.of(policyFromRules("permission",
                             frameworkConstraint(Map.of("Membership", "active"), "unknown-action", Operator.EQ, false)), "Policy with unknown actions types"),
-                    Arguments.of(policyFromRules("prohibition", namespace,
+                    Arguments.of(policyFromRules("prohibition",
                             frameworkConstraint(Map.of("Membership", "active"), "access", Operator.EQ, false)), "Access Policy with prohibition rule"),
-                    Arguments.of(policyFromRules("permission", namespace,
+                    Arguments.of(policyFromRules("permission",
                             frameworkConstraint(Map.of("UsagePurpose", "cx.core.industrycore:1"), "access", Operator.EQ, false)), "Access policy permission with not allowed constraints"),
-                    Arguments.of(policyFromRules("permission", namespace,
+                    Arguments.of(policyFromRules("permission",
                             frameworkConstraint(Map.of("BusinessPartnerNumber", "BPN0022232"), "use", Operator.EQ, false)), "Usage policy permission with not allowed constraints"),
-                    Arguments.of(policyFromRules("prohibition", namespace,
+                    Arguments.of(policyFromRules("prohibition",
                             frameworkConstraint(Map.of("AffiliatesRegion", "cx.region.europe:1"), "use", Operator.EQ, false)), "Usage policy prohibition with not allowed constraints"),
-                    Arguments.of(policyFromRules("obligation", namespace,
+                    Arguments.of(policyFromRules("obligation",
                             frameworkConstraint(Map.of("UsagePurpose", "cx.core.industrycore:1"), "use", Operator.EQ, false)), "Usage policy obligation with not allowed constraints"),
-                    Arguments.of(policyFromRules("permission", namespace,
+                    Arguments.of(policyFromRules("permission",
                             frameworkConstraint(Map.of("WarrantyDurationMonths", 3), "use", Operator.EQ, false),
                             frameworkConstraint(Map.of("WarrantyDefinition", "cx.warranty.contractEndDate:1"), "use", Operator.EQ, false)), "Policy with mutually exclusive constraints")
             );
         }
     }
 
-    private Response createPolicyDefinition(JsonObject policy) {
-        JsonObject requestBody = Json.createObjectBuilder().add("@context",
-                Json.createObjectBuilder().add("@vocab", "https://w3id.org/edc/v0.0.1/ns/")).add("@type", "PolicyDefinition").add("policy", policy).build();
-        return (Response) PROVIDER.baseManagementRequest().basePath("/v3").contentType(ContentType.JSON).body(requestBody).when().post("/policydefinitions", new Object[0]).then().extract();
+    private Response createPolicyDefinition(ManagementApiVersion apiVersion, JsonObject policy) {
+        JsonValue context;
+        switch (apiVersion) {
+            case V3 -> context = Json.createObjectBuilder()
+                    .add("@vocab", "https://w3id.org/edc/v0.0.1/ns/")
+                    .build();
+            case V4 -> context = Json.createValue(EDC_CONNECTOR_MANAGEMENT_CONTEXT_V2);
+            default -> context = null;
+        }
+
+        var requestBody = Json.createObjectBuilder()
+                .add(CONTEXT, context)
+                .add(TYPE, "PolicyDefinition")
+                .add("policy", policy)
+                .build();
+        return (Response) PROVIDER.baseManagementRequest()
+                .basePath("/%s".formatted(apiVersion.apiPath))
+                .contentType(ContentType.JSON)
+                .body(requestBody)
+                .when()
+                .post("/policydefinitions", new Object[0])
+                .then().extract();
     }
 
-    private static JsonObject policyFromRules(String ruleType, String policyDefinition, JsonObject... rules) {
+    private static JsonObject policyFromRules(String ruleType, JsonObject... rules) {
         var rulesArrayBuilder = Json.createArrayBuilder();
         for (JsonObject rule : rules) {
             rulesArrayBuilder.add(rule);
         }
         var contextArrayBuilder = Json.createArrayBuilder();
         contextArrayBuilder.add(ODRL_CONTEXT);
-        if (!policyDefinition.isBlank()) {
-            contextArrayBuilder.add(policyDefinition);
-        }
+        contextArrayBuilder.add(CX_POLICY_2025_09_CONTEXT);
 
         return Json.createObjectBuilder()
+                .add(ID, "id")
                 .add(CONTEXT, contextArrayBuilder)
                 .add(TYPE, "Set")
                 .add(ruleType, rulesArrayBuilder)
                 .build();
     }
 
-    private static JsonObject policyWithEmptyRule(String action, String policyContext) {
+    private static JsonObject policyWithEmptyRule(String action) {
         var rule = Json.createObjectBuilder()
                 .add("action", action)
                 .build();
@@ -278,12 +289,24 @@ public class PolicyDefinitionEndToEndTest {
         rulesArrayBuilder.add(rule);
         var contextArrayBuilder = Json.createArrayBuilder();
         contextArrayBuilder.add(ODRL_CONTEXT);
-        contextArrayBuilder.add(policyContext);
+        contextArrayBuilder.add(CX_POLICY_2025_09_CONTEXT);
 
         return Json.createObjectBuilder()
+                .add(ID, "id")
                 .add(CONTEXT, contextArrayBuilder)
                 .add(TYPE, "Set")
                 .add("permission", rulesArrayBuilder)
                 .build();
+    }
+
+    private enum ManagementApiVersion {
+        V3("v3"),
+        V4("v4");
+
+        final String apiPath;
+
+        ManagementApiVersion(String apiPath) {
+            this.apiPath = apiPath;
+        }
     }
 }
